REFACTORING WITH MINOR BEHAVIOR CHANGES
Separate the "data" objects from the algorithms (so that adding a new business capability does not impact the existing classes).
For example, we can now add XML parsing or serialization without any modification to the existing classes, as well as we can change the JSON library or the PDF library.

What was done :
1. Move the json parsing into json package inside a single JsonInvoice class (so that we can change our mind about json implementation without compromising the model, see the TODO about GSON)
2. Move the PDF printing into the PDF package inside a single PdfInvoice class (so that we do not bind our model to a PDF library)
3. Make the data objects immutables

Once that has been done, we added these minor changes :
1. Add options to the PDF printing (localization, translation, ...)
2. Change logo handling (using BufferedImage instead of a file path)
3. Change address to use fullname instead of title+last+first to be able to bill companies